Moved more shared helper methods into helper.js
